refactor: ファイル構造をリファクタリング
